requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=4.5.4
celery>=5.2.7
httpx>=0.24.0
bittensor>=6.0.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
sqlmodel>=0.0.8
pydantic-settings>=2.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-cov>=4.0.0
aiosqlite>=0.18.0
